refactor: add src folder
