+++ .travis.yml
@@ -2,6 +2,6 @@ language: node_js
 node_js:
   - 12
 script:
-  - npm run test:coverage
+  - npm run test:coverage -- --verbose
 after_script:
   - npm run coveralls
